fix: move files to outer folders
