+++ src/api/structures/connector/slack/ISlack.ts
@@ -1,7 +1,7 @@
 import { Placeholder, Prerequisite } from "@wrtnio/decorators";
 import { tags } from "typia";
 import { ContentMediaType } from "typia/lib/tags";
-import { StrictOmit } from "../../../../utils/strictOmit";
+import { StrictOmit } from "../../../../utils/types/strictOmit";
 import { ICommon } from "../common/ISecretValue";
 
 export namespace ISlack {
